+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `ProductBreadcrumb` component to render a product's breadcrumb structured data.
+- `SearchBreadcrumb` component to render a search's breadcrumb structured data.
 
 ## [0.4.0] - 2020-07-01
 ### Added